move controllers/ to shared/value
