+++ frontend/src/modules/alarms/limits/AlarmLimits.tsx
@@ -9,8 +9,8 @@ import { commitDraftRequest } from '../../../store/controller/actions';
 import { AlarmLimitsRequest, Range } from '../../../store/proto/mcu_pb';
 import { MessageType } from '../../../store/proto/types';
 import { setActiveRotaryReference } from '../../app/Service';
-import ValueClicker from '../../controllers/ValueClicker';
-import ValueSlider from '../../controllers/ValueSlider';
+import ValueClicker from '../../shared/value/ValueClicker';
+import ValueSlider from '../../shared/value/ValueSlider';
 import useRotaryReference from '../../shared/rotary/useRotaryReference';
 
 const useStyles = makeStyles((theme: Theme) => ({
