+++ frontend/src/modules/shared/value/ToggleValue.tsx
