+++ frontend/src/modules/shared/value/ValueClicker.tsx
@@ -7,8 +7,8 @@ import { Grid, Button, makeStyles, Theme } from '@material-ui/core';
 import KeyboardArrowUp from '@material-ui/icons/KeyboardArrowUp';
 import KeyboardArrowDown from '@material-ui/icons/KeyboardArrowDown';
 import { Subscription } from 'rxjs';
-import { setActiveRotaryReference, getActiveRotaryReference } from '../app/Service';
-import RotaryEncodeController from '../shared/rotary/RotaryEncodeController';
+import { setActiveRotaryReference, getActiveRotaryReference } from '../../app/Service';
+import RotaryEncodeController from '../rotary/RotaryEncodeController';
 
 const useStyles = makeStyles((theme: Theme) => ({
   root: {
