+++ frontend/src/modules/shared/value/ValueDisplay.tsx
