+++ frontend/src/modules/shared/value/ValueModal.tsx
@@ -4,8 +4,8 @@
 import React, { useCallback, useEffect, useRef } from 'react';
 import { makeStyles, Theme, Grid, Typography } from '@material-ui/core';
 import { shallowEqual, useSelector } from 'react-redux';
-import { getRotaryEncoder } from '../../store/controller/selectors/frontend_pb';
-import RotaryEncodeController from '../shared/rotary/RotaryEncodeController';
+import { getRotaryEncoder } from '../../../store/controller/selectors/frontend_pb';
+import RotaryEncodeController from '../rotary/RotaryEncodeController';
 import ValueClicker from './ValueClicker';
 
 const useStyles = makeStyles((theme: Theme) => ({
