+++ frontend/src/modules/shared/value/ValueSelectorDisplay.tsx
