+++ frontend/src/modules/shared/value/ValueSlider.tsx
