+++ frontend/src/modules/shared/value/ValueSpinner.tsx
